Adjust donationPage
