+++ .env
@@ -1,2 +1,2 @@
 CHOKIDAR_USEPOLLING=true
-REACT_APP_URL = https://adogta-backend-make-it-real.herokuapp.com
+REACT_APP_URL = http://localhost:8080
